Add cheat-sh-search
This could do with a bit more adding, perhaps better formatting of the
output (actually, I'm pondering the idea of a cheat-sh-mode for all output
so I can make jumping around cheat.sh easier) too. But there seems to be a
small issue with the non-highlighted output of cheat.sh right now (I've
reported it at https://github.com/chubin/cheat.sh/issues/7) so I'll leave
this until I know if that's a bug or intended behaviour.