:sparkles: Attempt to guess the topic for cheat-sh-guess-topic
The idea here is that, if possible, we try and guess the topic when using
`cheat-sh-search-topic`, based off the major-mode (to start with anyway).
Given that I don't really know the list of languages that cheat.sh knows
about, I've added the list as a `defcustom` which can easily be extended by
the user.

Done to explore options to help satisfy the question in #1.