:sparkles: Add some more major modes to the mode/topic map
I can't see an easy way of getting a list of known languages from cheat.sh
itself, so I'm having to do this bit-by-bit, which isn't brilliant. This is
pretty much why I think the major-mode of a language should be handled on
the server some where. There's a far small list of editors and IDEs than
there is languages that cheat.sh might understand, so adding support for
"oh, it's Emacs as the client and awk-mode means awk" feels like a much
nicer and easier solution.

See #1.